pyproject.toml
[tool.poetry]
name = "cyclops"
version = "0.1.1"
description = "Framework for healthcare ML implementation"
authors = ["Vector AI Engineering <cyclops@vectorinstitute.ai>"]
license = "Apache-2.0"
repository = "https://github.com/VectorInstitute/cyclops"
documentation = "https://vectorinstitute.github.io/cyclops/"

[tool.poetry.dependencies]
python = "3.9.7"
pandas = "^1.4.1"
SQLAlchemy = "^1.4.32"
ConfigArgParse = "^1.5.3"
python-dotenv = "^0.19.2"
matplotlib = "^3.5.1"
torch = "^1.11.0"
alibi-detect = "^0.9.1"
alibi = "^0.6.5"
shap = "^0.40.0"
xgboost = "^1.5.2"
pytest = "^7.1.1"
pre-commit = "^2.17.0"
pylint = "^2.12.2"
nb-black = "^1.0.7"
black = "^22.1.0"
tables = "^3.7.0"
Sphinx = "^4.4.0"
numpydoc = "^1.2"
sphinx-rtd-theme = "^1.0.0"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest-cov = "^3.0.0"
pyarrow = "^7.0.0"
sphinxcontrib-apidoc = "^0.3.0"
sphinx-autodoc-typehints = "^1.18.1"
llvmlite = "0.38.0"
psycopg2-binary = "^2.9.3"
plotly = "5.7.0"
pyparsing = "3.0.8"
dash = "^2.4.1"
dash-bootstrap-components = "^1.1.0"
Flask-Caching = "^1.10.1"
colorama = "^0.4.4"
prefect = "2.0b6"
seaborn = "^0.11.2"
myst-parser = "^0.18.0"
torchxrayvision = "^0.0.37"
dash-mantine-components = "^0.10.2"
dash-cool-components = "^0.1.8"
hydra-core = "^1.2.0"
dash-iconify = "^0.1.2"
dask = {extras = ["dataframe"], version = "^2022.9.1"}
sphinx-copybutton = "^0.5.0"

[tool.poetry.group.dev.dependencies]
evidently = "0.1.45.dev0"
mlflow = "^1.24.0"
jupyter = "^1.0.0"
jupyterlab = "^3.4.2"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
install_types = true
pretty = true
non_interactive = true

[tool.pylint]
    [tool.pylint.master]
    disable=["import-error", "duplicate-code"]

    [tool.pylint.format]
    max-line-length=88

    [tool.pylint.typecheck]
    generated-members=["numpy.*", "torch.*"]

    [tool.pylint.similarities]
    ignore-imports=["yes"]

[tool.flake8]
max-line-length = 88

[tool.coverage]
    [tool.coverage.run]
    source=["cyclops"]
    omit=["tests/*", "*__init__.py"]

[tool.nbqa.addopts]
flake8 = [
    "--max-line-length=88"
]

[tool.nbqa.exclude]
flake8 = "tutorials/mimiciv/query_api_developer.ipynb"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
